pyproject.toml
[tool.poetry]
name = "batchedit"
version = "0.1.0"
description = ""
authors = ["Justin Hewlett <hewlett.jus@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
Pillow = "^10.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
